refactor: Filter Component를 page 파일에서 분리
