+++ src/app/_components/GlobalNavigationBar.tsx
@@ -1,4 +1,4 @@
-import { Box, Typography } from '@mui/material';
+import { Box, Divider, Typography } from '@mui/material';
 import { PRODUCT_MENU } from '@/src/data/product';
 import { productStore, IProductState } from '@/src/store/productStore';
 
@@ -7,21 +7,25 @@ function GlobalNavigationBar() {
     productStore<IProductState>((state) => state);
 
   return (
-    <Box className="w-full h-12 flex items-center justify-center">
-      {PRODUCT_MENU.productType.map((type) => (
-        <Typography
-          onClick={() => setSelectProductType(type)}
-          className={`my-auto mx-5 ${
-            selectProductType === type
-              ? 'text-main underline underline-offset-8'
-              : ''
-          } hover:underline hover:underline-offset-8 hover:cursor-pointer hover:text-main transition-all ease-in-out`}
-          key={type}
-        >
-          {type}
-        </Typography>
-      ))}
-    </Box>
+    <>
+      <Box className="w-full h-12 flex items-center justify-center">
+        {PRODUCT_MENU.productType.map((type) => (
+          <Typography
+            onClick={() => setSelectProductType(type)}
+            className={`my-auto mx-5 ${
+              selectProductType === type
+                ? 'text-main underline underline-offset-8'
+                : ''
+            } hover:underline hover:underline-offset-8 hover:cursor-pointer hover:text-main transition-all ease-in-out`}
+            key={type}
+          >
+            {type}
+          </Typography>
+        ))}
+      </Box>
+
+      <Divider className="border-gray-400 mb-3" />
+    </>
   );
 }
