+++ src/app/_components/index.ts
@@ -1,4 +1,5 @@
 export { default as DisplayProducts } from './DisplayProducts';
+export { default as FilterChipsBox } from './FilterChipsBox';
 export { default as GlobalNavigationBar } from './GlobalNavigationBar';
 export { default as Header } from './Header';
 export { default as SideNavigationBar } from './SideNavigationBar';
