+++ src/components/ui/index.ts
@@ -1,5 +1,5 @@
 export { default as FilterChip } from './FilterChip';
-export { default as FilterList } from './FilterList';
+export { default as SideFilterList } from './SideFilterList';
 export { default as ProductColor } from './ProductColor';
 export { default as ProductInfo } from './ProductInfo';
 export { default as SkeletonUI } from './SkeletonUI';
